requirements.txt
GitPython>=3.1.0
pathspec>=0.9.0
alive-progress>=3.1.0